+++ src/modules/action-request/action-request.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { ActionRequestService } from './action-request.service';
+
+@Module({
+  providers: [ActionRequestService],
+  exports: [ActionRequestService],
+})
+export class ActionRequestModule {}